code/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
filelock==3.9.0
Flask==2.3.1
Flask-Cors==3.0.10
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
ratelimiter==1.2.0.post0
redis==4.5.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
tensorflow-metal==0.7.1
termcolor==2.2.0
textblob==0.17.1
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
tweepy==4.13.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.3.1
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
